third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/onerroreventhandler-expected.txt

This is a testharness.js-based test.
FAIL onerror + ErrorEvent + Window assert_equals: expected 5 but got 1
FAIL onerror + !ErrorEvent + Window assert_equals: expected 5 but got 1
PASS onerror + Document
Harness: the test ran to completion.
